# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Business Jarvis API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BusinessStatus = typeof BusinessStatus[keyof typeof BusinessStatus];


export const BusinessStatus = {
  active: 'active',
  inactive: 'inactive',
  pending: 'pending',
} as const;

export interface Business {
  id: number;
  name: string;
  city: string;
  country: string;
  lat: number;
  lng: number;
  industry: string;
  status: BusinessStatus;
  managerId: number;
  managerName: string;
  managerEmail: string;
  createdAt: string;
  /** @nullable */
  description?: string | null;
}

export interface BusinessInput {
  name: string;
  city: string;
  country: string;
  lat: number;
  lng: number;
  industry: string;
  managerName: string;
  managerEmail: string;
  description?: string;
}

export type BusinessUpdateStatus = typeof BusinessUpdateStatus[keyof typeof BusinessUpdateStatus];


export const BusinessUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
  pending: 'pending',
} as const;

export interface BusinessUpdate {
  name?: string;
  city?: string;
  country?: string;
  lat?: number;
  lng?: number;
  industry?: string;
  status?: BusinessUpdateStatus;
  description?: string;
}

export type ReportPeriod = typeof ReportPeriod[keyof typeof ReportPeriod];


export const ReportPeriod = {
  day: 'day',
  week: 'week',
  month: 'month',
} as const;

export interface Report {
  id: number;
  businessId: number;
  period: ReportPeriod;
  revenue: number;
  orders: number;
  profit: number;
  date: string;
  /** @nullable */
  notes?: string | null;
}

export type ReportInputPeriod = typeof ReportInputPeriod[keyof typeof ReportInputPeriod];


export const ReportInputPeriod = {
  day: 'day',
  week: 'week',
  month: 'month',
} as const;

export interface ReportInput {
  businessId: number;
  period: ReportInputPeriod;
  revenue: number;
  orders: number;
  profit: number;
  date: string;
  notes?: string;
}

export interface DashboardStats {
  totalRevenue: number;
  totalOrders: number;
  totalProfit: number;
  activeBusinesses: number;
  totalBusinesses: number;
  period: string;
}

export interface BusinessMetric {
  id: number;
  name: string;
  city: string;
  country: string;
  lat: number;
  lng: number;
  industry: string;
  revenue: number;
  orders: number;
  profit: number;
}

export interface AiSummary {
  period: string;
  summary: string;
  highlights: string[];
  generatedAt: string;
}

export interface ConnectInput {
  name: string;
  city: string;
  country: string;
  lat: number;
  lng: number;
  industry: string;
  managerName: string;
  managerEmail: string;
  accessCode: string;
  description?: string;
}

export type ListReportsParams = {
businessId: number;
period?: ListReportsPeriod;
};

export type ListReportsPeriod = typeof ListReportsPeriod[keyof typeof ListReportsPeriod];


export const ListReportsPeriod = {
  day: 'day',
  week: 'week',
  month: 'month',
} as const;

export type FetchLatestReportParams = {
businessId: number;
period?: FetchLatestReportPeriod;
};

export type FetchLatestReportPeriod = typeof FetchLatestReportPeriod[keyof typeof FetchLatestReportPeriod];


export const FetchLatestReportPeriod = {
  day: 'day',
  week: 'week',
  month: 'month',
} as const;

export type GetDashboardStatsParams = {
period?: GetDashboardStatsPeriod;
};

export type GetDashboardStatsPeriod = typeof GetDashboardStatsPeriod[keyof typeof GetDashboardStatsPeriod];


export const GetDashboardStatsPeriod = {
  day: 'day',
  week: 'week',
  month: 'month',
} as const;

export type GetTopBusinessesParams = {
period?: GetTopBusinessesPeriod;
limit?: number;
};

export type GetTopBusinessesPeriod = typeof GetTopBusinessesPeriod[keyof typeof GetTopBusinessesPeriod];


export const GetTopBusinessesPeriod = {
  day: 'day',
  week: 'week',
  month: 'month',
} as const;

export type GetAiSummaryParams = {
period?: GetAiSummaryPeriod;
};

export type GetAiSummaryPeriod = typeof GetAiSummaryPeriod[keyof typeof GetAiSummaryPeriod];


export const GetAiSummaryPeriod = {
  day: 'day',
  week: 'week',
  month: 'month',
} as const;