# lib/api-zod/src/generated/types/aiSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Business Jarvis API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiSummary {
  period: string;
  summary: string;
  highlights: string[];
  generatedAt: string;
}