# lib/api-zod/src/generated/types/business.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Business Jarvis API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessStatus } from './businessStatus';

export interface Business {
  id: number;
  name: string;
  city: string;
  country: string;
  lat: number;
  lng: number;
  industry: string;
  status: BusinessStatus;
  managerId: number;
  managerName: string;
  managerEmail: string;
  createdAt: string;
  /** @nullable */
  description?: string | null;
}