# lib/api-zod/src/generated/types/businessInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Business Jarvis API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessInput {
  name: string;
  city: string;
  country: string;
  lat: number;
  lng: number;
  industry: string;
  managerName: string;
  managerEmail: string;
  description?: string;
}