# lib/api-zod/src/generated/types/businessMetric.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Business Jarvis API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessMetric {
  id: number;
  name: string;
  city: string;
  country: string;
  lat: number;
  lng: number;
  industry: string;
  revenue: number;
  orders: number;
  profit: number;
}