# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Business Jarvis API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportInputPeriod } from './reportInputPeriod';

export interface ReportInput {
  businessId: number;
  period: ReportInputPeriod;
  revenue: number;
  orders: number;
  profit: number;
  date: string;
  notes?: string;
}